Initial working REST client for Stash, based on disco_eggs_template.
Contains a class with a main function that will either create or delete
a repository by name, for either a user or a project.  Requires a config
file, the structure of which is not documented at all, but which is where
we get the host name for the Stash server.