Guess project/repo/branch from local conditions, if asked.
Added code to read information from the local git repository and use it to
guess the source branch, project and repo name for pull requests.

Could also use this to guess the stash hostname, but then there are a lot
of things you could do if you started to fix the REST-client configuration.