Client refactor, step 3/many: switch to using client directly.
Made all the functions that were called by the CLI package into instance
methods of the StashRestClient object, and made the object responsible
for its own credential management, eliminating some weird coupling
between the rest client and the CLI package.

Also added a configuration option for remote username, and a command
line option for server hostname, to balance things out, and made
the CLI module marshall those and create a client object, instead
of setting constants in the rest package.